Considerar somente pendências ativas (not excluida) e organizando consultas de diário no repositório próprio
