+++ src/SME.SGP.Aplicacao/Queries/PendenciaDiarioBordo/ObterIdPendenciaDiarioBordoPorAulaEComponenteId/ObterIdPendenciaDiarioBordoPorAulaEComponenteIdQueryHandler.cs
@@ -1,9 +1,6 @@
 ﻿using MediatR;
-using SME.SGP.Dominio;
+using SME.SGP.Dominio.Interfaces;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
