+++ src/SME.SGP.Aplicacao/Queries/PendenciaDiarioBordo/ObterIdPendenciaDiarioBordoPorAulaId/ObterIdPendenciaDiarioBordoPorAulaIdQueryHandler.cs
@@ -1,10 +1,8 @@
 ﻿using MediatR;
-using SME.SGP.Dominio;
+using SME.SGP.Dominio.Interfaces;
 using SME.SGP.Infra;
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
