+++ src/SME.SGP.Aplicacao/Queries/PendenciaDiarioBordo/ObterPendenciasDiarioBordo/ObterPendenciasDiarioBordoQueryHandler.cs
@@ -1,10 +1,9 @@
 ﻿using MediatR;
 using SME.SGP.Dominio;
+using SME.SGP.Dominio.Interfaces;
 using SME.SGP.Infra;
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -12,9 +11,9 @@ namespace SME.SGP.Aplicacao
 {
     public class ObterPendenciasDiarioBordoQueryHandler : IRequestHandler<ObterPendenciasDiarioBordoQuery, IEnumerable<AulaComComponenteDto>>
     {
-        private readonly IRepositorioPendenciaDiarioBordo repositorioPendenciaDiarioBordo;
+        private readonly IRepositorioPendenciaDiarioBordoConsulta repositorioPendenciaDiarioBordo;
 
-        public ObterPendenciasDiarioBordoQueryHandler(IRepositorioPendenciaDiarioBordo repositorioPendenciaDiarioBordo)
+        public ObterPendenciasDiarioBordoQueryHandler(IRepositorioPendenciaDiarioBordoConsulta repositorioPendenciaDiarioBordo)
         {
             this.repositorioPendenciaDiarioBordo = repositorioPendenciaDiarioBordo ?? throw new ArgumentNullException(nameof(repositorioPendenciaDiarioBordo));
         }
