+++ src/SME.SGP.Aplicacao/Queries/PendenciaDiarioBordo/VerificaSeExistePendenciaDiarioComPendenciaId/VerificaSeExistePendenciaDiarioComPendenciaIdQueryHandler.cs
@@ -1,9 +1,5 @@
 ﻿using MediatR;
-using SME.SGP.Dominio;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+using SME.SGP.Dominio.Interfaces;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -12,9 +8,9 @@ namespace SME.SGP.Aplicacao
     public class VerificaSeExistePendenciaDiarioComPendenciaIdQueryHandler : IRequestHandler<VerificaSeExistePendenciaDiarioComPendenciaIdQuery, bool>
     {
         private readonly IMediator mediator;
-        private readonly IRepositorioPendenciaDiarioBordo repositorioPendenciaDiarioBordo;
+        private readonly IRepositorioPendenciaDiarioBordoConsulta repositorioPendenciaDiarioBordo;
 
-        public VerificaSeExistePendenciaDiarioComPendenciaIdQueryHandler(IMediator mediator, IRepositorioPendenciaDiarioBordo repositorioPendenciaDiarioBordo)
+        public VerificaSeExistePendenciaDiarioComPendenciaIdQueryHandler(IMediator mediator, IRepositorioPendenciaDiarioBordoConsulta repositorioPendenciaDiarioBordo)
         {
             this.mediator = mediator;
             this.repositorioPendenciaDiarioBordo = repositorioPendenciaDiarioBordo;
