+++ src/SME.SGP.Dados/Repositorios/RepositorioPendenciaAulaConsulta.cs
@@ -123,40 +123,6 @@ namespace SME.SGP.Dados.Repositorios
             }, commandTimeout: 60);
         }
 
-        public async Task<IEnumerable<PossuiPendenciaDiarioBordoDto>> TurmasPendenciaDiarioBordo(IEnumerable<long> aulasId, string turmaId, int bimestre)
-        {
-            var sqlQuery = new StringBuilder(@"select DISTINCT a.turma_id as TurmaId, a.aula_cj as AulaCJ
-                                                  from aula a
-                                                  inner join periodo_escolar pe on pe.tipo_calendario_id = a.tipo_calendario_id ");
-
-            sqlQuery.AppendLine(@" where a.data_aula between pe.periodo_inicio and pe.periodo_fim
-                                    and a.turma_id = @turmaId and pe.bimestre = @bimestre and a.id = ANY(@aulas) ");
-
-            return await database.Conexao.QueryAsync<PossuiPendenciaDiarioBordoDto>(sqlQuery.ToString(),
-               new
-               {
-                   turmaId,
-                   aulas = aulasId.ToArray(),
-               }, commandTimeout: 60);
-        }
-
-        public async Task<IEnumerable<long>> TrazerAulasComPendenciasDiarioBordo(string componenteCurricularId, string professorRf, bool ehGestor, string codigoTurma)
-        {
-            var disciplinaId = Convert.ToInt32(componenteCurricularId);
-            var sqlQuery = string.Empty;
-            if (ehGestor)
-            {
-                sqlQuery = @"select aula_id from pendencia_diario_bordo pdb join aula a on a.id = pdb.aula_id where a.turma_id = @codigoTurma";
-                return await database.Conexao.QueryAsync<long>(sqlQuery, new { codigoTurma }, commandTimeout: 60);
-            }
-            else
-            {
-                sqlQuery = @"select distinct pdb.aula_id from pendencia_diario_bordo pdb where pdb.componente_curricular_id = @disciplinaId and pdb.professor_rf = @professorRf";
-                return await database.Conexao.QueryAsync<long>(sqlQuery, new { professorRf, disciplinaId }, commandTimeout: 60);
-            }
-
-        }
-
         public async Task<IEnumerable<Aula>> ListarPendenciasAtividadeAvaliativa(long dreId, long ueId, int anoLetivo)
         {
             var sqlQuery = @"select distinct a.id, a.turma_id, a.disciplina_id, a.professor_rf,
@@ -237,18 +203,6 @@ namespace SME.SGP.Dados.Repositorios
 
             return await database.Conexao.QueryFirstOrDefaultAsync<Turma>(query, new { pendenciaId });
         }
-
-        public async Task<Turma> ObterTurmaPorPendenciaDiario(long pendenciaId)
-        {
-            var query = @"select t.* 
-                         from pendencia_diario_bordo pdb
-                        inner join aula a on a.id = pdb.aula_id 
-                        inner join turma t on t.turma_id = a.turma_id
-                        where pdb.pendencia_id = @pendenciaId ";
-
-            return await database.Conexao.QueryFirstOrDefaultAsync<Turma>(query, new { pendenciaId });
-        }
-
 
         public async Task<IEnumerable<PendenciaAulaDto>> ObterPendenciasAulasPorPendencia(long pendenciaId)
         {
@@ -462,34 +416,10 @@ namespace SME.SGP.Dados.Repositorios
                         join componente_curricular cc on cc.id = a.disciplina_id::int8 
                         where u.rf_codigo = @codigoRf and cc.id = @componenteCurricularId 
                         and pe.id = @periodoEscolarId and p.tipo = @tipoPendencia 
-                        and t.turma_id = @turmaCodigo and t.ue_id = @ueId
+                        and t.turma_id = @turmaCodigo and t.ue_id = @ueId and not p.excluido  
                         order by p.criado_em desc";
 
             return (await database.Conexao.QueryFirstOrDefaultAsync<long>(sql, new { componenteCurricularId, codigoRf, periodoEscolarId, tipoPendencia, turmaCodigo, ueId }));
-        }
-
-        public async Task<long> ObterPendenciaDiarioBordoPorComponenteProfessorPeriodoEscolar(long componenteCurricularId, string codigoRf, long periodoEscolarId)
-        {
-            try
-            {
-                var sql = @"select p.Id from pendencia p 
-                        join pendencia_diario_bordo pdb on pdb.pendencia_id = p.id 
-                        join pendencia_usuario pu on pu.pendencia_id = p.id 
-                        join usuario u on u.id = pu.usuario_id 
-                        join aula a on a.id = pdb.aula_id 
-                        join periodo_escolar pe on pe.tipo_calendario_id = a.tipo_calendario_id
-                        join componente_curricular cc on cc.id = pdb.componente_curricular_id
-                        where u.rf_codigo = @codigoRf and cc.id = @componenteCurricularId 
-                        and pe.id = @periodoEscolarId and p.tipo = @tipoPendencia 
-                        order by p.criado_em desc";
-
-                var retorno =  (await database.Conexao.QueryFirstOrDefaultAsync<long>(sql, new { componenteCurricularId, codigoRf, periodoEscolarId, tipoPendencia = (int)TipoPendencia.DiarioBordo }, commandTimeout: 60));
-                return retorno;
-            }
-            catch (Exception ex)
-            {
-                throw ex;
-            }
         }
 
         public async Task<long> ObterPendenciaPorDescricaoTipo(string descricao, TipoPendencia tipoPendencia)
