+++ src/SME.SGP.Dominio.Interfaces/Repositorios/IRepositorioPendenciaDiarioBordo.cs
@@ -1,9 +1,4 @@
 ﻿using SME.SGP.Dominio.Interfaces;
-using SME.SGP.Infra;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace SME.SGP.Dominio
@@ -12,7 +7,5 @@ namespace SME.SGP.Dominio
     {
         Task Excluir(long aulaId, long componenteCurricularId);
         Task ExcluirPorAulaId(long aulaId);
-        Task<bool> VerificarSeExistePendenciaDiarioComPendenciaId(long pendenciaId);
-        Task<IEnumerable<AulaComComponenteDto>> ListarPendenciasDiario(string turmaId, long[] componentesCurricularesId);
     }
 }
