+++ teste/SME.SGP.Aplicacao.Teste/Queries/Listao/ObterIndicativoPendenciasAulasPorTipoQueryHandlerTeste.cs
@@ -2,10 +2,7 @@
 using Moq;
 using SME.SGP.Dominio.Interfaces;
 using SME.SGP.Infra;
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Xunit;
@@ -16,14 +13,16 @@ namespace SME.SGP.Aplicacao.Teste.Queries
     {
 
         private readonly ObterIndicativoPendenciasAulasPorTipoQueryHandler query;
-        private readonly Mock<IRepositorioPendenciaAulaConsulta> repositorioPendenciaAula;
+        private readonly Mock<IRepositorioPendenciaAulaConsulta> repositorioPendenciaAulaConsulta;
+        private readonly Mock<IRepositorioPendenciaDiarioBordoConsulta> repositorioPendenciaDiarioBordoConsulta;
         private readonly Mock<IMediator> mediator;
 
         public ObterIndicativoPendenciasAulasPorTipoQueryHandlerTeste()
         {
-            repositorioPendenciaAula = new Mock<IRepositorioPendenciaAulaConsulta>();
+            repositorioPendenciaAulaConsulta = new Mock<IRepositorioPendenciaAulaConsulta>();
+            repositorioPendenciaDiarioBordoConsulta = new Mock<IRepositorioPendenciaDiarioBordoConsulta>();
             mediator = new Mock<IMediator>();
-            query = new ObterIndicativoPendenciasAulasPorTipoQueryHandler(repositorioPendenciaAula.Object);
+            query = new ObterIndicativoPendenciasAulasPorTipoQueryHandler(repositorioPendenciaAulaConsulta.Object,repositorioPendenciaDiarioBordoConsulta.Object);
         }
 
         [Fact]
@@ -42,10 +41,10 @@ namespace SME.SGP.Aplicacao.Teste.Queries
             var aulas = new List<long>();
             aulas.Add(123);
 
-            repositorioPendenciaAula.Setup(x => x.TrazerAulasComPendenciasDiarioBordo("512", "7941706", false, ""))
+            repositorioPendenciaDiarioBordoConsulta.Setup(x => x.TrazerAulasComPendenciasDiarioBordo("512", "7941706", false, ""))
                 .ReturnsAsync(aulas);
 
-            repositorioPendenciaAula.Setup(x => x.TurmasPendenciaDiarioBordo(aulas, "2386241", 1))
+            repositorioPendenciaDiarioBordoConsulta.Setup(x => x.TurmasPendenciaDiarioBordo(aulas, "2386241", 1))
                 .ReturnsAsync(dados);
 
             mediator.Setup(x => x.Send(It.IsAny<ObterIndicativoPendenciasAulasPorTipoQuery>(), It.IsAny<CancellationToken>()))
